Aida

+++ srcs/parsing/vec_utils.c
@@ -6,7 +6,7 @@
 /*   By: aneitenb <aneitenb@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 13:37:58 by aneitenb          #+#    #+#             */
-/*   Updated: 2024/05/21 13:57:31 by aneitenb         ###   ########.fr       */
+/*   Updated: 2024/06/13 19:54:43 by aneitenb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	vec_replace_special(t_vec *dst, void *src, size_t index)
 	str = *(char **)vec_get(dst, index);
 	if (vec_insert(dst, &src, index) < 0)
 		return (-1);
-	if (vec_remove(dst, (size_t)index + 1) < 0)
+	if (vec_remove(dst, index + 1) < 0)
 		return (-1);
 	free(str);
 	return (1);
